requirements.txt
chess==1.9.4
stockfish==3.28.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
logging>=0.5.1.2
